fix: circular imports & change task, category event_spaces
