+++ todo/__init__.py
@@ -1,2 +1,4 @@
 from .todo_category_sio import controller as category_events
 from .todo_task_rst import controller as task_namespace
+from .todo_task_sio import controller as task_events
+
