+++ todo/todo_category_sio.py
@@ -1,32 +1,13 @@
 from flask_fullstack import DuplexEvent, EventSpace, EventController
-from flask_socketio import join_room, leave_room
-from pydantic import BaseModel
 from common import User, CategoryTodo, db
+from todo.config import UserRoom
 
 controller: EventController = EventController()
 
 
 @controller.route()
-class CategoryEventSpace(EventSpace):
-
-    @classmethod
-    def room_name(cls, user_id: int) -> str:
-        return f"cat-{user_id}"
-
-    class UserIdModel(BaseModel):
-        user_id: int
-
-    @controller.argument_parser(UserIdModel)
-    @controller.force_ack()
-    def open_room(self, user: User.id) -> None:
-        join_room(self.room_name(user.id))
-
-    @controller.argument_parser(UserIdModel)
-    @controller.force_ack()
-    def close_room(self, user: User) -> None:
-        leave_room(self.room_name(user.id))
-
-    class CreateModel(CategoryTodo.CreationBaseModel, UserIdModel):
+class CategoryEventSpace(EventSpace, UserRoom):
+    class CreateModel(CategoryTodo.CreationBaseModel, UserRoom.UserIdModel):
         pass
 
     @controller.argument_parser(CreateModel)
@@ -42,7 +23,7 @@ class CategoryEventSpace(EventSpace):
         category = CategoryTodo.create(name=name, user_id=user.id)
         event.emit_convert(category, self.room_name(user.id))
 
-    class UpdateModel(CategoryTodo.CreationBaseModel, UserIdModel):
+    class UpdateModel(CategoryTodo.CreationBaseModel, UserRoom.UserIdModel):
         category_id: int
 
     @controller.argument_parser(UpdateModel)
@@ -64,7 +45,7 @@ class CategoryEventSpace(EventSpace):
         event.emit_convert(category, self.room_name(user.id))
         return category
 
-    class DeleteModel(UserIdModel):
+    class DeleteModel(UserRoom.UserIdModel):
         category_id: int
 
     @controller.argument_parser(DeleteModel)
