+++ todo/todo_task_rst.py
@@ -4,9 +4,8 @@ from flask import session
 from flask_restx import Resource
 from flask_fullstack import ResourceController
 from flask_restx.reqparse import RequestParser
-from datetime import datetime
 
-from common import TaskTodo as Task, User, get_datetime
+from common import TaskTodo as Task, User
 
 controller = ResourceController(name="tasks", path="/")
 
@@ -28,28 +27,9 @@ class TodoListTasks(Resource):
         return Task.get_all(session['user_id'])
 
 
-@controller.route("/create_task/")
-class TodoCreateTask(Resource):
-    @controller.jwt_authorizer(User, check_only=True)
-    @controller.argument_parser(task_parser)
-    @controller.marshal_with(Task.IndexModel)
-    def post(self, **kwargs):
-        if Task.find_first_by_kwargs(name=kwargs["name"], user_id=session['user_id']) is None:
-            return Task.create(
-                name=kwargs['name'],
-                description=kwargs['description'],
-                start_task=get_datetime(kwargs['start_task']),
-                end_task=get_datetime(kwargs['end_task']),
-                category_id=kwargs['category_id'],
-                user_id=session['user_id']
-            )
-        controller.abort(404, "task alredy exist")
-
-
 @controller.route("/detail_task/<string:task_name>")
 class TodoDetailTask(Resource):
     error_message = "task todo not found task_name incorrect"
-    success_message = "task was deleted"
 
     @controller.jwt_authorizer(User, check_only=True)
     @controller.marshal_with(Task.IndexModel)
@@ -57,18 +37,3 @@ class TodoDetailTask(Resource):
         if not (task := Task.find_first_by_kwargs(name=task_name, user_id=session['user_id'])) is None:
             return task
         controller.abort(404, self.error_message)
-
-    @controller.jwt_authorizer(User, check_only=True)
-    @controller.argument_parser(task_parser)
-    @controller.marshal_with(Task.IndexModel)
-    def patch(self, **kwargs) -> Task:
-        if (task := Task.find_first_by_kwargs(name=kwargs['task_name'], user_id=session['user_id'])) is None:
-            controller.abort(404, self.error_message)
-        return Task.change_values(task, **kwargs)
-
-    @controller.jwt_authorizer(User, check_only=True)
-    def delete(self, task_name: str):
-        if (task := Task.find_first_by_kwargs(name=task_name, user_id=session['user_id'])) is None:
-            controller.abort(404, self.error_message)
-        task.delete()
-        return "task was deleted"
